requirements.txt
Django==2.1.7
django-redis==4.10.0
djangorestframework==3.9.2
pytz==2018.9
redis==3.2.0
